Install py3 depends if py3 exists #247
This will install both Python2 and Python3 depends if the system has both installed, so that users don't get an error if they try to use the python3 when the code installs for python2.

If python3 doesn't exist, it won't try to install it.  If python2 doesn't exist, it won't try to install 2.  If the default is 3, it will install just 3.  Only should do both if 2 is default, but 3 exists.