Improved Cloud Support (#354)
This updates the memcached API to bmemcached, which supports the faster binary protocol and SASL auth used by cloud providers. It maintains seamless backward compatibility for python-memcached. It also adds a new step that parses environmental variables after it reads the configuration file. Command line argument will take priority.

Some notes, memcached_port was removed as it wasn't actually used in the code and it makes sense to just include it as part of hostname:port as more than one memcache server can be defined via comma separation. Also, libmemcached-dev was removed as I don't think it is actually needed by python-memcached or python-binary-memcached as they're both fully python apis. I think it's only needed by pylibmc as that's a C implementation. memcached was removed from the Docker file as it is defined in the docker compose, so it downloads as a separate container.

Updated the docker-compose.yml and also added code for deploying the Demo box, as it uses the Environment variable to enable it.