Identicons #335 (#355)
Identicon avatars if no user or team avatars are supplied in the avatar directory by the admin. This just replaces the default avatar image, which was the same for every user. Based on player handle or team name. The admin can use the old default image by just dragging it into the user or team directory. The identicon is only used when no other avatar images are supplied.