Fix #384
If you use the User Management to change the a user from player to admin, it didn't remove them from the team (in the case of player mode - the team of 1).  Some listings, like the scoreboard and user management, filter out players with no team.  Since it didn't remove the team, they stayed listed on the scoreboard.