Game levels (#447)
Adds additional options to the game levels and management. You can now choose to unlock a level via points reached or set it to "managed" where you control who is added to a level. A new button "Edit Access" is on the level which allows you to add and remove users. You can now set the type on level 0, so you can change it from default to prevent new users from automatically getting the level. Note that managed levels are hidden from users until they are added to the level.