585 flagdepends (#611)
* use async in edit flag

The cause of this bug was a race condition as both getBoxFlags and getDetails were executed at once.  In some cases the flag details would return before the boxflags returned and updated the flaglist, resulting in it selecting the empty option.

This fix makes getBoxFlags return a promise which we now await before moving to getDetails.

* fix flag create jquery error

The create flag views aren't modals, but the links where a box.uuid is passed into them used data-toggle="modal" which caused a jquery error.

Removed data-toggle="modal" for these button links to fix.