validate image size
Fixes an issue where users might upload an image under the minimum dimensions (500x250) during registration. This would leave their user account in an unusable state, and completely unmanageable even from the UI as an admin.

This fix validates the image size as part of the validation check, raising a validation error if the image dimensions are too small.